geoviewer-streamlit-python/pyproject.toml
[project]
name = "geoviewer-streamlit-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "email-validator>=2.2.0",
    "geopandas>=1.0.1",
    "leafmap>=0.42.13",
    "owslib>=0.33.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.39",
    "streamlit>=1.44.0",
    "streamlit-extras>=0.6.0",
]
